fix(fx-lib): guard the RTPatch limit table against signed overflow
The fuzz smoke found a UBSan signed-integer-overflow in build_limits
(rtpatch.cpp), the adaptive-Huffman limit-table builder: on crafted patch
headers `s3 = s4 * 2` doubles the working value every level while
`s4 = s3 - r16s(...)` feeds it back, so s4 grows without bound until
`s4 * 16` no longer fits an int. It is harmless on real patches — their
limits always fit 16 bits — but undefined behaviour on hostile input, and
the `fx patch` path runs this over untrusted bytes.

s3 and s4 mirror the original codec's 16-bit registers, which wrap at 16
bits; the reconstruction modelled them with a plain int and never wrapped.
Every value the routine produces is stored mod 2^16 via w16 and the
recurrence is linear, so reducing s3/s4 mod 2^16 each step (the same s16 /
&0xFFFF idiom the rest of the decoder already uses) is identical on every
valid patch while bounding s4 to the int16 range — the *2/*4/*16 products
can no longer overflow. The byte-identical RTPatch tests are unchanged.

The CI crash input is added as a regression seed under
fuzz/corpus/fuzz_rtpatch/ (the harness had no committed corpus, which is
why the 60s smoke fuzzed from empty and reached this in the first place).

Note for the future: aggressive `-fork` fuzzing also surfaced a SEGV that
symbolises into libFuzzer's own fork-management runtime, does not
reproduce for any single input, and never recurs under single-process or
`-jobs` fuzzing (14 workers x 300s, zero crashes) — i.e. a fork-mode
artifact, not an rtpatch defect. The single-process 60s smoke CI runs is
clean after this fix.

+++ lib/src/rtpatch.cpp
@@ -177,11 +177,19 @@ private:
     void build_limits(int start) {
         uint32_t gc = r32(0x0c), lim = r32(0x10);
         int num = r16u(0x04);
-        int s3 = (start == 0) ? 2 : r16s(lim + (start - 1) * 8) * 2;
+        // s3/s4 mirror the original codec's 16-bit registers, so every step wraps
+        // to 16 bits — the same modular model the rest of this decoder uses (s16,
+        // &0xFFFF). Without the wrap, `s3 = s4 * 2` doubles each level and s4 grows
+        // without bound on crafted input until `s4 * 16` overflows int (UBSan
+        // signed-overflow). Everything here is stored mod 2^16 via w16 and the
+        // recurrence is linear, so wrapping is identical on every valid patch while
+        // bounding s4 to the int16 range — the *2/*4/*16 products can no longer
+        // overflow. (Regression seed: fuzz/corpus/fuzz_rtpatch/overflow-build_limits.)
+        int s3 = (start == 0) ? 2 : s16((uint16_t)(r16s(lim + (start - 1) * 8) * 2));
         for (int level = start; level < num; ++level) {
-            int s4 = s3 - r16s(gc + level * 2);
+            int s4 = s16((uint16_t)(s3 - r16s(gc + level * 2)));
             w16(lim + level * 8, s4);
-            s3 = s4 * 2;
+            s3 = s16((uint16_t)(s4 * 2));
             w16(lim + level * 8 + 2, s3);
             w16(lim + level * 8 + 4, s4 * 4);
             w16(lim + level * 8 + 6, s4 * 16);
